project/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
tensorflow==2.13.0
transformers==4.33.2
gunicorn==21.2.0
numpy==1.24.3
requests==2.31.0
torch==2.0.1
protobuf==4.24.3